メインパネルをコンポーネント化
