+++ packages/editor/src/main/browser/components/launcher/ProjectList.tsx
@@ -1,16 +1,42 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import {Project} from "../../../ts/common/Preference";
 import ProjectItem from "./ProjectItem";
+import ProjectRepository from "../../repository/ProjectRepository";
 
 interface Props {
-
-  projectList: Project[]
-
+  onProjectLoaded?: (projectList: Project[]) => void
 }
 
-const ProjectList: React.FunctionComponent<Props> = (props) => {
+const ProjectList: React.FunctionComponent<Props> = ({onProjectLoaded}) => {
 
-  const items = props.projectList.map(p => {
+  const [projectList, setProjectList] = useState<Project[]>([]);
+
+  useEffect(() => {
+
+    let unmounted = false;
+
+    function handleListChange(list: Project[]) {
+      if (unmounted) {
+        return;
+      }
+      setProjectList(list);
+      if (onProjectLoaded) {
+        onProjectLoaded(list);
+      }
+    }
+
+    const projectList = ProjectRepository.instance().getProjectList();
+
+    projectList.on(handleListChange);
+
+    return function cleanup() {
+      unmounted = true;
+      projectList.off(handleListChange);
+    }
+
+  }, [onProjectLoaded, setProjectList]);
+
+  const items = projectList.map(p => {
     return <ProjectItem project={p}/>;
   });
