+++ packages/editor/src/main/browser/components/organism/LauncherMainPanel/LauncherMainPanel.css
@@ -1,0 +1,2 @@
+.main {
+}
